Feat : Running Response 수정

+++ Backend/Running/src/main/java/com/ygmg/running/entity/RunningDetail.java
@@ -29,7 +29,7 @@ public class RunningDetail {
 
     private Time runningTime;
 
-    private Long runningKcal;
+    private Double runningKcal;
 
     private Double runningDistance;
